database set up

+++ CardTypes/Spell.cs
@@ -13,11 +13,11 @@ namespace MTCG_Server.CardTypes
             _guID = guID;
             _name = name;
             if (name.Substring(0, 5) == "Water")
-                _elementID = (int)Server.ElementID.Water;
+                _elementID = Server.ElementID.Water;
             else if (name.Substring(0, 4) == "Fire")
-                _elementID = (int)Server.ElementID.Fire;
+                _elementID = Server.ElementID.Fire;
             else
-                _elementID = (int)Server.ElementID.Normal;
+                _elementID = Server.ElementID.Normal;
             _dmg = dmg;
             _type = "Spell";
             //Console.WriteLine("Spell is created");
